+++ goDB/Fakes/FakeResult.php
@@ -7,7 +7,6 @@ namespace go\DB\Fakes;
 
 use go\DB\Helpers\Fetchers\Arr;
 
-
 /**
  * @author Oleg Grigoriev <go.vasac@gmail.com>
  */